refactor: 가게 별 미션 조회 api 도메인 수정
