+++ nunu/week7/src/modules/missions/controllers/mission.controller.ts
@@ -1,10 +1,8 @@
 import {
   Controller,
-  Get,
   Patch,
   Path,
   Post,
-  Query,
   Route,
   SuccessResponse,
   Tags,
@@ -13,7 +11,6 @@ import { StatusCodes } from "http-status-codes";
 import {
   completeUserMission,
   createChallangeMission,
-  getMissions,
 } from "../services/mission.service.js";
 import { AppError } from "../../../common/errors.js";
 import { success } from "../../../common/responses.js";
@@ -46,35 +43,6 @@ export class MissionController extends Controller {
     this.setStatus(StatusCodes.CREATED);
 
     return success(mission, "도전 미션이 추가되었습니다.", StatusCodes.CREATED);
-  }
-
-  @Get("stores/{storeId}/missions")
-  public async handleGetMissions(
-    @Path() storeId: number,
-    @Query() cursor: number = 0,
-    @Query() limit: number = 5,
-  ) {
-    console.log("미션 목록 조회를 요청했습니다!");
-
-    // storeId 검증
-    if (!storeId || Number.isNaN(storeId)) {
-      throw new AppError(
-        "유효하지 않은 storeId 입니다.",
-        StatusCodes.BAD_REQUEST,
-      );
-    }
-
-    // cursor, limit 검증
-    if (Number.isNaN(cursor) || Number.isNaN(limit) || limit < 1) {
-      throw new AppError(
-        "유효하지 않은 cursor 또는 limit 입니다.",
-        StatusCodes.BAD_REQUEST,
-      );
-    }
-
-    const result = await getMissions(storeId, cursor, limit);
-
-    return success(result, "미션 목록 조회를 성공했습니다.", StatusCodes.OK);
   }
 
   @Patch("{missionId}/complete")
