+++ nunu/week7/src/modules/missions/dtos/mission.dto.ts
@@ -14,40 +14,6 @@ export const responseFromUserMission = (
   };
 };
 
-export interface GetMissionResponse {
-  id: number;
-  storeName: string;
-  category: string;
-  description: string;
-  rewardPoint: number;
-  expireDate: string;
-}
-
-export interface GetMissionsResult {
-  missions: GetMissionResponse[];
-  totalPages: number;
-  hasNext: boolean;
-}
-
-export const responseFromMissions = (
-  missions: any[],
-  hasNext: boolean,
-  totalPages: number,
-): GetMissionsResult => {
-  return {
-    missions: missions.map((mission) => ({
-      id: Number(mission.id),
-      storeName: mission.store.name,
-      category: mission.store.storeType,
-      description: mission.description,
-      rewardPoint: Number(mission.rewardPoint),
-      expireDate: mission.expireDate!.toISOString().split("T")[0],
-    })),
-    hasNext,
-    totalPages,
-  };
-};
-
 export interface CompleteUserMissionResponse {
   missionId: number;
   isCompleted: boolean;
