+++ nunu/week7/src/modules/stores/repositories/store.repository.ts
@@ -24,3 +24,52 @@ export const getRegionById = async (regionId: number) => {
     where: { id: regionId },
   });
 };
+
+export const getMissionsByStore = async (
+  storeId: number,
+  cursor: number,
+  limit: number,
+) => {
+  const missions = await prisma.mission.findMany({
+    where: {
+      storeId,
+      ...(cursor && {
+        id: {
+          gt: cursor,
+        },
+      }),
+    },
+    select: {
+      id: true,
+      description: true,
+      rewardPoint: true,
+      expireDate: true,
+      store: {
+        select: {
+          name: true,
+          storeType: true,
+        },
+      },
+    },
+    orderBy: {
+      id: "asc",
+    },
+    take: limit + 1,
+  });
+
+  const hasNext = missions.length > limit;
+
+  return {
+    missions: hasNext ? missions.slice(0, limit) : missions,
+    hasNext,
+  };
+};
+
+// totalPages 계산
+export const countmissionsByStore = async (storeId: number) => {
+  return await prisma.mission.count({
+    where: {
+      storeId,
+    },
+  });
+};
