--- nunu/week7/src/routes/mission.route.ts
@@ -1,12 +1,0 @@
-import { Router } from "express";
-import {
-  handleChallengeMission,
-  handleCompleteMission,
-} from "../modules/missions/controllers/mission.controller.js";
-
-const router = Router();
-
-router.post("/:missionId/challenges", handleChallengeMission);
-router.patch("/:missionId/complete", handleCompleteMission);
-
-export default router;
