--- nunu/week7/src/routes/user.route.ts
@@ -1,7 +1,0 @@
-import { Router } from "express";
-import { RegisterRoutes } from "../generated/routes.js";
-
-const router = Router();
-RegisterRoutes(router);
-
-export default router;
